backend/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# PyTorch and Deep Learning
torch==2.1.0
torchvision==0.16.0
torchaudio==0.16.0
transformers==4.36.1
accelerate==0.25.0
safetensors==0.4.1

# Audio and Video Processing
librosa==0.10.0
soundfile==0.12.1
ffmpeg-python==0.2.1
moviepy==1.0.3
opencv-python==4.8.1.78
pillow==10.1.0

# Utilities
numpy==1.24.3
scipy==1.11.4
tqdm==4.66.1
requests==2.31.0

# Development and Testing
pytest==7.4.3
black==23.12.0
isort==5.13.2
flake8==6.1.0
